+++ src/app/site-edit/site-edit.component.spec.ts
@@ -622,6 +622,8 @@ describe('SiteEditComponent', () => {
     expect(component.addFileType).toEqual("New");
     expect(component.siteFileForm.get("file_date").value).not.toEqual(null);
     expect(component.siteFileForm.get("photo_date").value).not.toEqual(null);
+    expect(component.siteFileForm.get("source_id").value).not.toEqual(null);
+    expect(component.siteFileForm.get("agency_id").value).not.toEqual(null);
   });
 
   it('should show alert and stop loading if site form is invalid', () => {