+++ src/app/site-edit/site-edit.component.ts
@@ -552,9 +552,17 @@ export class SiteEditComponent implements OnInit {
     this.siteFileForm.controls["photo_date"].setValue(this.siteFiles.FileEntity.photo_date);
     
     // Set source name and agency automatically
+    // Member id
     let member_id = JSON.parse(localStorage.getItem('currentUser')).member_id;
     this.siteFiles.FileEntity.source_id = member_id;
     this.siteFileForm.controls['source_id'].setValue(member_id);
+    // FULLname
+    this.siteFiles.FileEntity.FULLname = JSON.parse(localStorage.getItem('currentUser')).fname + " " +  JSON.parse(localStorage.getItem('currentUser')).lname;
+    this.siteFileForm.controls['FULLname'].setValue(this.siteFiles.FileEntity.FULLname);
+    // Agency
+    this.siteFiles.FileEntity.agency_id = JSON.parse(localStorage.getItem('currentUser')).agency_id;
+    this.agencyNameForCap = JSON.parse(localStorage.getItem('currentUser')).agency_name;
+    this.siteFileForm.controls['agency_id'].setValue(this.siteFiles.FileEntity.agency_id);
   }
 
   getFileTypeSelection(event) {